fixup! drm/asahi: Initial buffer manager implementation

+++ drivers/gpu/drm/asahi/buffer.rs
@@ -380,7 +380,7 @@ impl Buffer::ver {
             raw.page_count.store(page_count, Ordering::Relaxed);
             raw.block_count.store(new_count as u32, Ordering::Relaxed);
             raw.last_page
-                .store((page_count - 1) as u32, Ordering::Relaxed);
+                .store(page_count - 1, Ordering::Relaxed);
         });
 
         Ok(true)
